requirements.txt
google-generativeai
python-dotenv
unidecode
cloudscraper
beautifulsoup4
requests
pandas
fastapi
SQLAlchemy